rename: 폴더명 수정, 폴더 이동 - ToggleButton 공통 컴포넌트로 사용하여 ui 폴더로 이동 - 폴더 이동으로 인한 경로 수정 - constant 폴더 확장성 고려하여 constants 폴더로 수정
