+++ src/components/ui/ToggleButton.jsx
@@ -23,7 +23,7 @@ const ToggleButton = ({ size = "m" }) => {
   const { buttonSize, thumbSize, activeTranslate } = toggleSize[size];
 
   return (
-    <button
+    <Button
       className={`relative cursor-pointer rounded-full transition-colors duration-200 ${
         isToggled ? "bg-black" : "bg-neutral-300"
       } ${buttonSize}`}
@@ -34,7 +34,7 @@ const ToggleButton = ({ size = "m" }) => {
           isToggled ? activeTranslate : "translate-x-0"
         } ${thumbSize}`}
       />
-    </button>
+    </Button>
   );
 };
